rest-api/pyproject.toml
[project]
name = "rest-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles==25.1.0",
    "annotated-types==0.7.0",
    "anyio==4.11.0",
    "click==8.3.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.119.0",
    "h11==0.16.0",
    "idna==3.10",
    "python-jose[cryptography]>=3.3.0",
    "passlib[argon2]>=1.7.4",
    "pydantic[email]==2.12.0",
    "pydantic-core==2.41.1",
    "python-multipart==0.0.20",
    "slowapi>=0.1.9",
    "sniffio==1.3.1",
    "starlette==0.48.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn==0.37.0",
]
